tbdev turndown: delete unneeded util code
